REFACTOR - Histoy imports for conversion model
- Se arreglaron los imports para el modelo de conversiones
